datapath: Move rev-DNAT NodePort xlation to to-netdev
Previously, the reverse DNAT translation for a reply from a local
NodePort endpoint was done only in bpf_lxc via the
CILIUM_CALL_IPV{4,6}_NODEPORT_REVNAT tail call.

This was breaking a case when a NodePort endpoint was running in the
host netns (bpf_lxc is not attached to endpoint in the host netns).
Therefore, a reply from such endpoint was never rev-DNAT xlated.

Fix this by performing the xlation from bpf_netdev's to-netdev (egress).

Signed-off-by: Martynas Pumputis <m@lambda.lt>