test/K8sServices: test IPv4 datagram fragments support
Add a test to check that fragmented datagrams are correctly routed
through the load balancer and the conntrack lookup.

The test consists in sending fragmented datagrams with netcat from one
"client" pod to a "server" endpoint. To avoid changing the policy in use
for K8sServices, we reuse the existing UDP echo server deployed on the
DS test nodes (although we do not send valid TFTP commands). Reception
of the datagrams is validated by searching the conntrack table for the
correct number of packets for the relevant flow. Using ClusterIP or
NodePort, a number of IP addresses and ports are used to target the
endpoint and validate that the datagram is received in various cases.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>