bpf: add support for IPv4 fragmentation to CT/policy lookup
Similarly to L4 port extraction for the IPv4 load balancer, add
fragments support for policy checks.

Because the tuple used for L4 (or L3+L4) policy checks is performed as
part of the conntrack lookup, call for L4 ports extraction in
`ct_lookup4()`. There are some implications:

- Fragments are not supported for policy checks if `CONNTRACK` is not
  defined.

- Fragment processing is currently performed each time `ct_lookup4()` is
  called for a TCP/UDP packet. This is the case for `ipv4_policy()`,
  which we primarily target, but also for `handle_ipv4_from_lxc()`,
  `(rev_)nodeport_lb4()`, NAT functions, or `lb4_local()`. Not sure if
  this is desirable.

If fragments should not be processed on some paths we could deactivate
that with a bool value passed down to `ct_lookup4()`. Alternatively, we
can also take port extraction out of there and move it to e.g.
`ipv4_policy()`, but we would need a flag to prevent `ct_lookup4()` from
overwriting the collected values. Or we could refactor L4 ports
extraction out of `ct_lookup4()`, but that may require more work to make
sure that we keep consistency on all paths.

The need to check whether the fragment is the first for that datagram
also implies that we need to revalidate header data to access the IP
header for each call to `ct_lookup4()`. This could maybe be optimised by
looking at the `ip4->id` field before pointers are invalidated, and
passing it down to the function.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>