agent: add several new flags to control the Cilium's datapath events exposure
This commit introduces three new configuration flags for the Cilium agent, allowing users to choose the bpf event types they want to expose to Cilium monitor and Hubble.
- `--enable-drop-notify`: Enables `DropNotification` by default.
- `--enable-policy-verdict-notify`: Enables `PolicyVerdictNotification` by default.
- `--enable-trace-notify`: Enables `TraceNotification` by default.

The default values for these flags remain set to `true`, not changing the current behaviour.

In our case, we found particularly useful to disable the TraceNotification in order to reduce the CPU overhead on some of our nodes when Hubble is enabled as we were mostly interested into dropped packets.

Signed-off-by: Maxime Visonneau <maxime.visonneau@gmail.com>